move progress tracker and merge conf
